feat: :sparkles: implement controllers and configure Swagger for API documentation
This commit introduces the initial set of controllers for managing API endpoints in the Laravel application. Each controller is responsible for handling requests related to core functionalities, such as user management and task operations.   Additionally, Swagger has been installed and configured to provide interactive API documentation. This setup includes annotations in the controllers that describe the endpoints, request parameters, and responses. The integration of Swagger enhances the developer experience by offering clear, up-to-date documentation for the API, making it easier to test and understand the endpoints. This feature is essential for maintaining a robust and user-friendly API.

documentacion api